Fix formatting, merge issues, and enum class change (ATTACK_TYPE)

+++ src/map/entities/mobentity.cpp
@@ -803,7 +803,7 @@ void CMobEntity::DistributeRewards()
                 {
                     RoeDatagramList datagrams;
                     datagrams.push_back(RoeDatagram("mob", (CMobEntity*)this));
-                    datagrams.push_back(RoeDatagram("atkType", this->BattleHistory.lastHitTaken_atkType));
+                    datagrams.push_back(RoeDatagram("atkType", static_cast<uint8>(this->BattleHistory.lastHitTaken_atkType)));
                     roeutils::event(ROE_MOBKILL, (CCharEntity*)PMember, datagrams);
                 }
             });
